### 2018-01-30 #190 @evanodell
nomisr
_Downloads UK official statistics from the Nomis API and converts to R objects, primarily tibbles. The API is based around different statistical geographies._

```
Package: nomisr
Title: Access Nomis UK Labour Market Data with R
Version: 0.0.2
Authors@R: person("Evan Odell", email="evanodell91@gmail.com",
  role=c("aut", "cre"),
  comment = c(ORCID='0000-0003-1845-808X'))
Author: Evan Odell [aut, cre]
Maintainer: Evan Odell <evanodell91@gmail.com>
Description: Access UK official statistics from the Nomis database through R.
    Nomis includes data from the Census, the Labour Force Survey, DWP benefit
    statistics and other economic and demographic data from the Office for
    National Statistics, based around statistical geographies. See
    <https://www.nomisweb.co.uk/api/v01/help> for full API documentation.
  curl
```

https://github.com/EvanOdell/nomisr

[e.g., "data extraction, because the package parses a scientific data file format"]

_Data retrieval, because the package assists the downloading of a data from an API into R._

_Demographers, economists, geographers, public health researchers, any other social scientists who are interested in geographic factors. The package will aid reproducibility, reduce the need to manually download area profiles, and allow easy linking of different datasets covering the same geographic area._

_There are no other R packages specifically for retrieving this data. Some of the data on Nomis is also available through the UK Data Service with the [`ukds`](https://cran.r-project.org/package=ukds) package, but other ONS data, including labour force, benefits, mortality and business statistics are not otherwise available through an R package. The UK Data Service also requires users registration, while Nomis does not, and does not, as far as I can tell, provide the same linking with statistical geographies as Nomis._

	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.
