### 2018-01-30 #190 @evanodell
Hi all, submitting this as a presubmission because I'm not sure how much it fits with rOpenSci's remit.

Best,

Evan
