### 2018-01-30 c729fbc
update vignette
