### 2018-01-30 #190 @karthik
# nomisr
### Editor checks:

---

#### Editor comments

Here are results of a check on the good practice package on some best practices. You can re-run this yourself (as you fix issues) by running `goodpractice::gp()` on your package. Please fix relevant issues before reviewers begin to take a look at your package.

```
── GP nomisr ───────────────────────────────────────────────────────────────────

It is good practice to

    R/data_download.R:103:NA
    R/data_download.R:137:NA
    R/data_download.R:139:NA
    R/data_download.R:141:NA
    R/data_download.R:143:NA
    ... and 35 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.

    R/data_download.R:118:1
    R/data_download.R:135:1
    R/data_download.R:157:1
    R/nomis_codes.R:52:1
    R/search.R:29:1
    ... and 3 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/data_download.R:143:14

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
```

---

Reviewers: Currently seeking reviewers
Due date:
