### 2018-01-31 e615337
adding more test coverages, tweaking some code for best practices
