+++ R/utils-query.R
@@ -4,7 +4,8 @@
 
 nomis_query_util <- function(query){
   
-  q <- jsonlite::fromJSON(paste0("https://www.nomisweb.co.uk/api/v01/dataset", query), flatten = TRUE)
+  q <- jsonlite::fromJSON(paste0("https://www.nomisweb.co.uk/api/v01/dataset",
+                                 query), flatten = TRUE)
   
   df <- tibble::as_tibble(q$structure$keyfamilies$keyfamily)
   
