+++ R/nomis_codes.R
@@ -8,14 +8,14 @@
 #' 
 #' Specifying `code` will return all the options for a given variable.
 #'
-#' @param id The ID of the particular dataset. If both \code{id} and 
-#' \code{code} are left empty, returns all available codes for all datasets.
+#' @param id The ID of the particular dataset. If both `id` and 
+#' `code` are left empty, returns all available codes for all datasets.
 #' @param code A string with the variable code to return options for. If left 
 #' empty, returns all the variables for the dataset specified by `id`. 
 #' Codes are not case sensitive. Defaults to `null`.
 #' @param type A string with options for a particular code value, to return 
 #' types of variables available for a given code. Defaults to `null`.
-#' @return A list of options.
+#' @return A tibble with options.
 #' @export
 #'
 #' @examples \dontrun{
@@ -49,7 +49,8 @@ nomis_codes <- function(id, code=NULL, type=NULL){
   type_query <- dplyr::if_else(is.null(type)==FALSE, paste0("/",type), "")
     
   a <- jsonlite::fromJSON(paste0("https://www.nomisweb.co.uk/api/v01/dataset/",
-                                 id,"/",code, type_query, "/def.sdmx.json?"), flatten=TRUE)
+                                 id,"/",code, type_query, "/def.sdmx.json?"),
+                          flatten=TRUE)
                      
   x <- as.data.frame(a$structure$codelists$codelist$code)
   
