+++ R/utils-collect.R
@@ -5,7 +5,9 @@
 
 nomis_collect_util <- function(query){
   
-  df <- suppressMessages(readr::read_csv(paste0("https://www.nomisweb.co.uk/api/v01/dataset", query)))
+  df <- suppressMessages(readr::read_csv(
+    paste0("https://www.nomisweb.co.uk/api/v01/dataset", query)
+    ))
   
   df
   
