+++ tests/testthat/test_search.R
@@ -9,4 +9,10 @@ test_that("nomis_search return expected format", {
   expect_type(y, "list")
   expect_true(tibble::is_tibble(y))
   
+  x <- nomis_search("seekers")
+  
+  expect_length(x, 14)
+  expect_type(x, "list")
+  expect_true(tibble::is_tibble(x))
+  
 })
