+++ R/search.R
@@ -7,8 +7,8 @@
 #'
 #' @param search A string to search for in the name and description of 
 #' datasets.
-#' @param keywords If \code{TRUE}, searches in dataset keywords instead of 
-#' names and descriptions. Defaults to \code{FALSE}. 
+#' @param keywords If `TRUE`, searches in dataset keywords instead of 
+#' names and descriptions. Defaults to `FALSE`. 
 #'
 #' @return A tibble with details on all datasets matching the search query.
 #' @export
@@ -25,8 +25,10 @@
 nomis_search <- function(search, keywords=FALSE){
   
   query <- dplyr::if_else(keywords==FALSE,
-                           paste0("/def.sdmx.json?search=*", search, "*"),
-                           paste0("/def.sdmx.json?search=keywords-", search, "*")
+                           paste0("/def.sdmx.json?search=*", 
+                                  search, "*"),
+                           paste0("/def.sdmx.json?search=keywords-", 
+                                  search, "*")
   
   )
                           
