### 2018-01-31 f9ebe92
update to pass all good practice checks
