+++ vignettes/introduction.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Introduction to Nomis
+# Introduction to the Nomis API
 
 Nomis provides free access to up-to-date official statistics from the Office for National Statistics (ONS).
 
