+++ DESCRIPTION
@@ -24,5 +24,6 @@ RoxygenNote: 6.0.1
 Suggests: 
   knitr,
   rmarkdown,
-  testthat
+  testthat,
+  curl
 VignetteBuilder: knitr
