### 2018-02-04 453db2e
update vignette yaml to include vignette title in index entry and toc
