### 2018-02-06 979ccd7
documentation improvements
