+++ R/data_download.R
@@ -1,22 +1,23 @@
 
 
-#' Retrieve nomis datasets
+#' Retrieve Nomis datasets
 #' 
 #' Retrieves specific datasets from nomis, based on their ID. To find dataset 
-#' IDs, use `\link{nomis_data_info}`. Datasets are retrived in csv format 
-#' and parsed with the `read_csv` function from the `readr` package. 
+#' IDs, use \code{\link{nomis_data_info}}. Datasets are retrived in csv format 
+#' and parsed with the \code{read_csv} function from the \code{readr} package. 
 #' 
-#' To find the code options for a given dataset, use `\link{nomis_codes}`.
+#' To find the code options for a given dataset, use \code{\link{nomis_codes}}.
 #' 
 #' This can be a very slow process if calling significant amounts of data.
 #' 
-#' Note the difference between the `time` and `date` parameters. The `time` 
-#' and `date` parameters should not be used at the same time. If they are, 
-#' the function will retrieve data based on the the `date` parameter. If given
-#' more than one query, `time` will return all data available between those 
-#' queries, inclusively, while `date` will only return data for the exact 
-#' queries specified. So `time=c("first","latest")` will return all data,
-#' while `date=c("first","latest")` will return only the first and latest
+#' Note the difference between the \code{time} and \code{date} parameters. 
+#' The \code{time} and \code{date} parameters should not be used at the same 
+#' time. If they are, the function will retrieve data based on the the 
+#' \code{date} parameter. If given more than one query, \code{time} will 
+#' return all data available between those queries, inclusively, while 
+#' \code{date} will only return data for the exact queries specified. So 
+#' \code{time=c("first","latest")} will return all data, while 
+#' \code{date=c("first","latest")} will return only the first and latest 
 #' data published.
 #' 
 #'
@@ -24,56 +25,56 @@
 #' @param time Parameter for selecting dates and date ranges. There are two 
 #' styles of values that can be used to query time. 
 #' 
-#' The first is one or more of `"latest"` (returns the latest available data), 
-#' `"previous"` (the date prior to `"latest"`), `"prevyear"` (the date one 
-#' year prior to `"latest"`) or `"first"` (the oldest available data for the 
-#' dataset). 
+#' The first is one or more of \code{"latest"} (returns the latest available 
+#' data), \code{"previous"} (the date prior to \code{"latest"}), 
+#' \code{"prevyear"} (the date one year prior to \code{"latest"}) or 
+#' \code{"first"} (the oldest available data for the dataset). 
 #' 
 #' The second style is to use or a specific date or multiple dates, in the 
 #' style of the time variable codelist, which can be found using the 
-#' `\link{nomis_codes}` function.
+#' \code{\link{nomis_codes}} function.
 #' 
-#' Values for the `time` and `date` parameters should not be used at the 
-#' same time. If they are, the function will retrieve data based on the 
-#' the `date` parameter.
+#' Values for the \code{time} and \code{date} parameters should not be used 
+#' at the same time. If they are, the function will retrieve data based 
+#' on the the \code{date} parameter.
 #' 
-#' Defaults to `NULL`.
+#' Defaults to \code{NULL}.
 #' 
 #' @param date Parameter for selecting specific dates. There are two styles 
 #' of values that can be used to query time. 
 #' 
-#' The first is one or more of `"latest"` (returns the latest available data), 
-#' `"previous"` (the date prior to `"latest"`), `"prevyear"` (the date one 
-#' year prior to `"latest"`) or `"first"` (the oldest available data for the 
-#' dataset). 
[…]
