### 2018-02-10 #190 @cderv
## Package Review for `nomisr`

Onboarding submission : https://github.com/ropensci/onboarding/issues/190

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

First of all, great package ! Very well done, Everything is good!
Well done really ! There was still medium coverage at the submission, it is 100% now and all run perfectly locally. Know that `goodpractice::gp` praises you right :

> ♥ Whee! Grand package! Keep up the fabulous work!

Everything is based on tibble class that makes it perfect to work inside the `tidyverse`. It is really great! However, maybe this could be more explained for those who are not used to tibbles and tidyverse ecosystem.
I find it kind of necessary because you also used list-columns in tibbles. This is not something easy for some newcomers to handles. Even for some _dplyr_ users who do not used that a lot, it is not something easy. Adding some example in the vignettes to show how to retrieve embedded data.frame inside some tibble's list column could be a good idea. To illustrate:
``` r
library(nomisr)
library(dplyr, warn.conflicts = F)
y <- nomis_data_info("NM_893_1")
y$annotations.annotation %>% class()
#> [1] "list"
y$annotations.annotation[[1]] %>% class()
#> [1] "data.frame"
y %>% pull(annotations.annotation) %>% class()
#> [1] "list"
y %>% pull(annotations.annotation)%>% .[[1]] %>% class()
#> [1] "data.frame"
y %>% pull(annotations.annotation) %>% purrr::pluck(1) %>% class()
#> [1] "data.frame"
y %>% pull(annotations.annotation) %>% purrr::map(1) %>% class()
#> [1] "list"
y %>% tidyr::unnest(annotations.annotation) %>% glimpse()
#> Observations: 14
#> Variables: 11
#> $ agencyid                             <chr> "NOMIS", "NOMIS", "NOMIS"...
#> $ id                                   <chr> "NM_893_1", "NM_893_1", "...
#> $ uri                                  <chr> "Nm-893d1", "Nm-893d1", "...
#> $ version                              <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1...
#> $ components.primarymeasure.conceptref <chr> "OBS_VALUE", "OBS_VALUE",...
#> $ components.timedimension.codelist    <chr> "CL_893_1_TIME", "CL_893_...
#> $ components.timedimension.conceptref  <chr> "TIME", "TIME", "TIME", "...
#> $ name.value                           <chr> "LC4408EW - Tenure by num...
#> $ name.lang                            <chr> "en", "en", "en", "en", "...
#> $ annotationtext                       <chr> "Current (being actively ...
#> $ annotationtitle                      <chr> "Status", "Keywords", "Un...
```

Following are my review comments, for improvement:

#### About Community Guidelines

+ There is no contribution guidelines in the README or other CONTRIBUTING.md.
+ DESCRIPTION needs some improvement : no need to use *Author* and *Maintainer* fields if you are using *Authors@R*. You can just use the latest. You can find some good example in [R package book](http://r-pkgs.had.co.nz/description.html#author).

#### Improvement for the [packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)

see below parts for details

+ Adding a code of conduct
+ Improve README considering it is the first entry point for discovering NOMIS API
+ Improve DESCRIPTION for authorship
+ [minor] follow [versionning advice](http://r-pkgs.had.co.nz/description.html#version) with `*.*.*.9000 ` for in development version
+ [question] I did not find why you need curl package in suggest...

#### README improvement

+ I would be more precise on what the NOMIS database is and for who it is. For example, I found that this submission issue mentioned UK database but not the README. You help file for the package as pretty clear description of what it is for. While reviewing, I had a doubt and went to check on [NOMIS](https://www.nomisweb.co.uk/). This is why I think the statement of need is not complete.
+ It feels like the README could get more examples on what your package can do. Only one of your functions is shown in the readme.
+ Put package documentation link in the title of the repo like in the [`reprex` repo](https://github.com/tidyverse/reprex). By the way, you can add a more descriptive description on your github repo and some tags.
+  could precise that it uses the tibble format (from the tidyverse)

#### Vignettes improvement

+ Vignette was not built when I installed the package with `devtools::install_github`. Did not know if this intended behavior or not.
+ I tried to build the vignette locally then: it worked.
    + However, there is a problem on result formatting with the tibble. (I think because you use `result = 'asis'` in code chunks) and because every results are not tables (`y$annotations.annotation` is a list for example.) so it does not print nicely.
    + The vignette do not show any code to get the results - I would show them for someone to learn how to use your function, as examples. (with `echo = TRUE`)
    + The formatting issue is visible on your [website](https://docs.evanodell.com/nomisr/articles/introduction.html) too
+ I think the vignette contains necessary information but does not show enough examples on how to retrieve data in the objects you return. You mentioned the three list columns with nested data.frame but not show how correctly extracts the info.
+ It is pretty clear that you package allows to retrieve a large amount of data from the NOMIS API. it seems necessary to always use `nomis_data_info`, then `nomis_codes` then `nomis_get_data` to be sure to not retrieve to much. The vignette helps you understand that. It could appears to in your readme, and it could be more clear how to construct a pipeline for good practice with this API.

These improvement are the reason why I did not check the vignettes box, even if there is one in the github repo and doc.
If you need help on all this, do not hesitate to ask.

#### About documentation of functions and examples

+ All functions are well documented.
+ I encountered some issues running the examples because of a timeout. I believe this is what you meant in your by "there are some limits to the amount of data that can be retrieved within a certain period of time, although those are not published.". The maybe some improvement to do on this. (see under)
+ Is this the reason why you enclosed each example in `\dontrun{}` statement ? am I not sure on what is the recommendation for API package like this one for example ?  `\dontrun{}` does allow anyone to run examples with something like `example("nomis_codes", package = "nomisr")`. Idea : Maybe `\donttest{}` is better here: does show in help and runs with `example`, but not run by `R CMD check` (so not run by CRAN ?) - see [rd vignette from roxygen2](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd.html)
+ I ran all tests with `devtools::run_examples(run = FALSE)` (in several time, due to timeout constraint). Everything is working. It works well for almost all, but your example assign systematically to a variable (named x or y - room for improvement) but never print to show head of results. I feel that is missing.
+ I said "almost all" because there are some very long request in `nomis_get_data`. The warning that you put in _Details_ states that it can be very slow process if calling significant amount of data. I would emphasize on this, saying this warning in _Description_ part of the help file. And these example should definitely not be ran.
+ I saw you made a pkgdown website. This is great !! and nice theme!
I did not find the source code of this website. I would have thought it leaved with the package source code. But no docs folder or gh-pages. I find it useful to have access for letting people PR your some typos or other thing and all the site from your package source code. However, you may have good reason.

#### Various comments on the code

+ [minor] I would put the url of the API in its own variable. I think it is better practice and more efficient to maintain in case of change in the url (e.g *V0.2*). The url is used several time and a change would mean copy pasting. another utils function could help building the url (`paste0` is used three times), with the same base url `def.sdmx.json`). Know that it exists some tools to help build some strings with parameters like `glue`, and specific one to handle url like `urltools` or those in `httr`. You may not need them here but it is worth knowing.

+ `suppressMessage` is used to hide the message of `readr::read_csv`. However, the message appears because column types are guessed and not set. It is good practice to set column type in `readr` with `column_types = cols(<something>)`. In an API package like this one, it make even more sense because it could helps be aware of any change in the format. (using `stop_for_problems()` or just see a message appear) - see [readr vignettes](http://readr.tidyverse.org/articles/readr.html#column-specification)
+ [minor] I don't know what are the good practice on this one. I just feel that you have a dplyr dependency in import for three functions `bind_rows`, `case_when` and `if_else` than can be easily replace with `rbind` or `if.else`. I feel that, if the former are very useful interactively in an analysis, using them in a package can be unnecessary dependency to a "big" package that can cause you some effort later on if anything change in dplyr. More of a personal point of view on this one.
+ However, it relates on how you are build your query url
```r
  query <- dplyr::if_else(keywords==FALSE,
                           paste0("/def.sdmx.json?search=*",
                                  search, "*"),
                           paste0("/def.sdmx.json?search=keywords-",
                                  search, "*")
```
could be rewritten without `dplyr::if_else`
```r
base_query <- "/def.sdmx.json?search="
if (keyword) {
    keyword_term <- "keywords-")
    keyword_term <- "*"
}
# or keyword_term <- if.else(keyword, "keywords-", "*")
query <- paste0(base_query, key_word_term, search, "*")
```
It would suppres a non essential dependency.
+ The style in code could be improve in some places. You can follow some style guide to help you.
like the one in ROpensci Guidelines linking to [Advanced R](http://r-pkgs.had.co.nz/r.html#style) or more recent the one used in the [tidyverse styleguide](http://style.tidyverse.org/). They even build some tools to help like [styler](http://styler.r-lib.org/) and [lintr](https://github.com/jimhester/lintr). Example of improvement:
   + put spaces after `for`or `if` and also around `=`, `==`, `+`, `>`, ...
   + indentation of code (in RStudio CTRL+i i useful). for example in `if () { } else { }`, lines of code that are inside brackets should be indented.
   + use `{}` with `if` when it does not fit on one line
   + More importantly, be consistent in all your script with the guideline you choose.
+ Know that when you want to stop on a condition, there is also `stopifnot`. For user experience, you can also return clearer error sometimes with `stop("API request did not return any results", call. = FALSE)` to not show the call for example.
+ A good practice is also to use meaningful name inside your internal code, even for temporary variable. There are a lot of `x`, `a`, `q`, ... This will improve readability and the ease for others to collaborate on your code (including future you)

#### About the timeout and size of data

+ If session is `interactive()`, it may be a good idea to warn clearly and even ask if we want to continue when lots of data are being requested. I find myself playing with the example trying to retrieve 208888 of your pages
```
Retrieving additional pages 1 of 208888
Retrieving additional pages 2 of 208888
```
I don't know if some other API packages are doing something to help user or if it is all the user responsibility to know what is he doing. (users will surely be already NOMIS data users).

+ As another improvement, I think you should try to create custom error message for the timeout issue as you know it is an issue that will impact user experience but cannot really prevent. I encountered:
```r
a <- nomis_codes("NM_1_1")
# Error in open.connection(con, "rb") :
#  Timeout was reached: Resolving timed out after 10000 milliseconds
```
You could use some `tryCatch` mechanism (some useful info in [Advanced R](http://adv-r.had.co.nz/Exceptions-Debugging.html#condition-handling)).

+ [thoughts] You may need more control on the request mechanism to deal with timeout. First, it seems for the website one need to contact NOMIS help-desk to know more about limitation. Then , your are relying on readr connection mechanism, and may need to use httr, crul or curl to better handle request to API with the specificity of NOMIS

#### About SMDX format on NOMIS

+ Do you know about the `rsdmx` package ? it is on [github](https://github.com/opensdmx/rsdmx) and CRAN, there is [WIKI](https://github.com/opensdmx/rsdmx/wiki). I did not find NOMIS in their listing, so I did not know if it works for you and how it could overlap with your package.
`rsdmx` is definitely more a generic package than yours. I just wanted to point it to you in case you did not know.

### Last words

Thank you very much letting me review this package. It is pretty simple one, yet powerful as it brings to R users lots of new open data for UK. And it does it well allowing to work with tidyverse tools directly thanks to the tibble format.

All my remarks were toward improvement. The package in its currents state is functional, is well tested and as the minimal documentation to get start with it.

@evanodell, if you have some questions or you want some help / advice on anything, feel free to reach to me, I will be glad to help you.

@karthik, I hope I am doing fine for my first review with Ropensci, please tell me if miss something and need to review further.
