+++ DESCRIPTION
@@ -1,12 +1,10 @@
 Package: nomisr
 Type: Package
 Title: Access Nomis UK Labour Market Data with R
-Version: 0.0.2
+Version: 0.0.2.9000
 Authors@R: person("Evan Odell", email="evanodell91@gmail.com",
   role=c("aut", "cre"),
   comment = c(ORCID='0000-0003-1845-808X'))
-Author: Evan Odell [aut, cre]
-Maintainer: Evan Odell <evanodell91@gmail.com>
 Description: Access UK official statistics from the Nomis database through R. 
     Nomis includes data from the Census, the Labour Force Survey, DWP benefit 
     statistics and other economic and demographic data from the Office for 
@@ -24,6 +22,5 @@ RoxygenNote: 6.0.1
 Suggests: 
   knitr,
   rmarkdown,
-  testthat,
-  curl
+  testthat
 VignetteBuilder: knitr
