### 2018-02-12 1d7effa
travis & gitignore updates
