+++ .travis.yml
@@ -10,6 +10,7 @@ cache: packages
 before_install:
   - sudo apt-get update
   - sudo apt-get install libcurl4-openssl-dev -y
+  - Rscript -e "install.packages('curl')"
 
 r_packages:
   - devtools
