### 2018-02-12 f1c4b95
travis tweaks
