### 2018-02-12 b3e1cca
switch travis to devel R
