### 2018-02-12 ec0e2a2
possible way around curl problems?
