### 2018-02-12 43cca15
install libcurl4 on travis
