+++ .travis.yml
@@ -9,8 +9,9 @@ cache: packages
 
 before_install:
   - sudo apt-get update
-  - sudo apt-get install libcurl4-openssl-dev
-  - Rscript -e "install.packages('curl')"
+  - sudo apt-get install -y libcurl-dev
+#  - Rscript -e "install.packages('curl')"
+# - sudo apt-get install libcurl4-openssl-dev
 
 r_packages:
   - devtools
