+++ .travis.yml
@@ -9,7 +9,7 @@ cache: packages
 
 before_install:
   - sudo apt-get update
-  - sudo apt-get install -y libcurl-dev
+  - sudo apt-get install -y libcurl4-openssl-dev 
 #  - Rscript -e "install.packages('curl')"
 # - sudo apt-get install libcurl4-openssl-dev
 
