+++ DESCRIPTION
@@ -22,5 +22,6 @@ RoxygenNote: 6.0.1
 Suggests: 
   knitr,
   rmarkdown,
-  testthat
+  testthat,
+  curl
 VignetteBuilder: knitr
