### 2018-02-12 6e4abe0
some improvements to docs and internals
