+++ R/utils-collect.R
@@ -5,8 +5,10 @@
 
 nomis_collect_util <- function(query) {
     
-    df <- suppressMessages(readr::read_csv(paste0("https://www.nomisweb.co.uk/api/v01/dataset", 
-        query)))
+    df <- readr::read_csv(
+      paste0(base_url, query), 
+      col_types = readr::cols(.default = readr::col_character())
+      )
     
     df
     
