+++ R/data_download.R
@@ -16,12 +16,13 @@
 #' \code{date} parameter. If given more than one query, \code{time} will 
 #' return all data available between those queries, inclusively, while 
 #' \code{date} will only return data for the exact queries specified. So 
-#' \code{time=c("first","latest")} will return all data, while 
-#' \code{date=c("first","latest")} will return only the first and latest 
+#' \code{time=c("first", "latest")} will return all data, while 
+#' \code{date=c("first", "latest")} will return only the first and latest 
 #' data published.
 #' 
 #'
 #' @param id The ID of the dataset to retrieve.
+#' 
 #' @param time Parameter for selecting dates and date ranges. There are two 
 #' styles of values that can be used to query time. 
 #' 
@@ -61,22 +62,34 @@
 #' @param geography The code of the geographic area to return data for. If 
 #' \code{NULL}, returns data for all available geographic areas, subject to 
 #' other parameters. Defaults to \code{NULL}.
+#' 
 #' @param measures The code for the statistical measure(s) to include in the 
 #' data. Accepts a single string or number, or a list of strings or numbers. 
 #' If \code{NULL}, returns data for all available statistical measures subject 
 #' to other parameters. Defaults to \code{NULL}.
+#' 
 #' @param sex The code for sexes included in the dataset. Accepts a string or 
-#' number, or a vector of strings or numbers. \code{7} will return results for 
-#' males and females, \code{6 }only females and \code{5} only males. 
+#' number, or a vector of strings or numbers. \code{nomisr} automatically voids 
+#' any queries for sex if it is not an available code for the 
+#' requested dataset.
+#' 
+#' There are two different codings used for sex, depending on the dataset. For 
+#' datasets using \code{"SEX"}, \code{7} will return results for 
+#' males and females, \code{6} only females and \code{5} only males. 
 #' Defaults to \code{NULL}, equivalent to \code{c(5,6,7)} for datasets where 
-#' sex is an option.
-#' @param exclude_missing If \code{TRUE}, excludes all missing values. 
-#' Defaults to \code{FALSE}.
+#' sex is an option. For datasets using \code{"C_SEX"}, \code{0} will return 
+#' results for males and females, \code{1} only males and 
+#' \code{2} only females. 
+#' 
 #' @param additional_queries Any other additional queries to pass to the API.
 #' See \url{https://www.nomisweb.co.uk/api/v01/help} for instructions on 
 #' query structure. Defaults to \code{NULL}.
+#' 
+#' @param exclude_missing If \code{TRUE}, excludes all missing values. 
+#' Defaults to \code{FALSE}.
 #'
-#' @return A tibble containing the selected dataset.
+#' @return A tibble containing the selected dataset. 
+#' By default, all tibble columns are parsed as characters.
 #' @export
 #' @seealso nomis_data_info
 #' @seealso nomis_codes
@@ -88,83 +101,124 @@
 #' y <- nomis_get_data(id="NM_1_1", time="latest")
 #' 
 #' # Return data for each country
-#' z <- nomis_get_data(id="NM_1_1", time="latest", geography="TYPE499", 
+#' z <- nomis_get_data(id="NM_1_1", time="latest", geography = "TYPE499",
 #'                     measures=c(20100, 20201), sex=5)
 #' 
 #' # Return data for Wigan
-#' a <- nomis_get_data(id="NM_1_1", time="latest", geography="1879048226", 
-#'                     measures=c(20100, 20201), sex=5)
+#' a <- nomis_get_data(id="NM_1_1", time="latest", geography = "1879048226",
+#'                     measures=c(20100, 20201), sex="5")
+#'                     
+#'                     
+#' b <- nomis_get_data(id="NM_168_1", time="latest", geography = "1879048226",
[…]
