+++ R/utils-base_urls.R
@@ -1,0 +1,6 @@
+
+
+
+base_url <- "https://www.nomisweb.co.uk/api/v01/dataset/"
+
+#devtools::use_data(base_url, internal = TRUE, overwrite = TRUE)
