+++ tests/testthat/test_search.R
@@ -3,7 +3,7 @@ context("nomis_search")
 
 test_that("nomis_search return expected format", {
   
-  y <- nomis_search("Claimants", keywords=TRUE)
+  y <- nomis_search("Claimants", keywords = TRUE)
   
   expect_length(y, 14)
   expect_type(y, "list")
