### 2018-02-12 0475beb
changed some examples around on get_data
