+++ R/data_download.R
@@ -95,24 +95,33 @@
 #' @seealso nomis_codes
 #' 
 #' @examples \dontrun{
-#' 
 #' x <- nomis_get_data(id="NM_1_1")
 #' 
 #' y <- nomis_get_data(id="NM_1_1", time="latest")
+#' }
+#' 
+#' \donttest{
 #' 
 #' # Return data for each country
 #' z <- nomis_get_data(id="NM_1_1", time="latest", geography = "TYPE499",
 #'                     measures=c(20100, 20201), sex=5)
+#'                     
+#' head(z)                     
 #' 
 #' # Return data for Wigan
 #' a <- nomis_get_data(id="NM_1_1", time="latest", geography = "1879048226",
 #'                     measures=c(20100, 20201), sex="5")
 #'                     
-#'                     
+#' head(a)
+#' 
+#'                                           
 #' b <- nomis_get_data(id="NM_168_1", time="latest", geography = "1879048226",
-#'                     measures=c(20100, 20201), sex="5")
+#'                     measures=c(20100, 20201), sex="0")
+#' 
+#' head(b)
 #' 
 #' }
+
 
 
 nomis_get_data <- function(id, time = NULL, date = NULL, geography = NULL,
