### 2018-02-12 bb9780d
update tests to improve coverage
