### 2018-02-13 1bcae74
fix potential record offset bug in nomis_get_data
