### 2018-02-19 e989829
additions to vignette
