### 2018-02-22 65d194b
add tidyr to suggests
