+++ DESCRIPTION
@@ -24,5 +24,6 @@ Suggests:
   rmarkdown,
   testthat,
   curl,
-  purrr
+  purrr,
+  tidyr
 VignetteBuilder: knitr
