### 2018-02-23 #190 @pegeler
## Package Review for `nomisr`

Onboarding submission: ropensci/onboarding#190

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Excellent work! This is a fine example of the level of quality and professionalism that I have
come to expect from rOpenSci contributions. This package is a great addition to the
complement of R functionality and provides a facile portal to a useful data source.

Many of the comments that I had have been addressed as a result of @cderv's prompt and thorough review.
In addition, development has continued to be very active over the past few weeks! Issues are
fixed as quickly as I find them! Below are a few more comments that I would like to add.

#### Basic Package Structure Overview

##### README.md

The README.md does not contain citation information, per rOpenSci packaging guide.
I am not sure how necessary that is since there is a CITATION file in the _`inst/`_ directory,
however, it should be an easy add.

I would concur with @cderv in suggesting additional usage examples.

##### DESCRIPTION

- The DESCRIPTION file does not currently specify a minimum version of R. I would
  recommend doing so. A minor release is preferred over a patch release, e.g.

  is better than

- I am not seeing where `curl` is used even though it is in the `Suggests` list.

##### CITATION

The CITATION file uses `meta$Date` to dynamically fill in date for the citation.
However, the `Date` entry was removed from the DESCRIPTION file per `gp` suggestion so
the date is NULL.

Since it is considered good practice to omit the `Date` entry because it may be prone to
falling out of date, it seems that hard-coding the date in the citation would be silly.
I briefly considered the use of a different field to dynamically update the citation date,
such as `Date/Publication` or `Packaged`. However, those will only work for users who are downloading
from a CRAN repos. Therefore I'm a little bit at a loss as to what the most elegant solution might be.
I might suggest breaking with good practice here and adding the `Date` field back in
since there is a reason to do so.
Suggestions are welcome.

The CITATION file should also end in a newline character :)

##### Documentation

- The man page for the package is named *nmisr*, not *nomisr*. This appears to be a typo.
- `@seealso` can link to the other package functions using `\link{}`.
- It appears that some descriptions may be intended to be multi-paragraph. Without explicitly
  using the `@description` and `@details` tags, the third code block is assumed to be the 'details'
  block by default. Therefore, using explicit roxygen tags might be helpful here.
- Nomis data comes packaged in very complex data structures. Please don't be afraid of 'over-explaining'.
  The more information in the documentation and vignettes, the better for the user, especially if they do
  not have prior experience with Nomis data.
- 'Nomis' in the title for `nomis_search()` should be capitalized to be consistent with the rest of the package.

#### Examples

- The example for `nomis_get_data` failed due to lack of data. I recommend finding a specific
  date range that will return results and then hard-code that into the example.
- Athough `\donttest{}` used to be perfect for API packages because it would not be run by
  `R CMD check` but would be run by `example()` and would be included in the man page, [it appears
  that the functionality has changed](http://r-pkgs.had.co.nz/man.html#man-functions).
  Now, according to Hadley's book, it appears that code enclosed
  in `\donttest{}` is indeed checked (though I have yet to confirm). Therefore,
  for long-running API calls, it might make sense to change back to `\dontrun{}` if you are
  concerned about time-out. I recall
  reading some best practice advice on CRAN's website at some point, but cannot find the reference
  anymore.
- The examples often show multiple ways to use the function, which is excellent. However,
  it might be a nice enhancement to add a little extra code to glimpse the data for the users to get
  a better feel for what they are getting out of the API call. This is especially helpful since
  the data that is returned can differ greatly depending on what args are passed.

- I like the switch to `echo = TRUE` for vignette code chunks. This is the user's opportunity
  to see the code in action for the first time. Including the code used to generate the results can
  aid in understanding and readability.
- Vignettes are not currently built. This is not a big deal. However, it would
  be a convenience for those installing from GitHub or local repos. This is especially true since
  the API calls can cause vignettes to take a long time to build. (Currently I cannot build vignettes
  because I keep getting hung up).
- The addition of the theme and TOC is a nice touch.
- It looks like the vignette might still be a little incomplete. I see that the final paragraph of
  *Querying data variables* is not finished.
- It's unnecessary to hard-code the results in code chunks. The results will be output when the vignette is knit.

- `nomis_search()`
    - [suggestion] Currently, users can either search based on generic search terms or search for specific keywords.
      Looking at the API docs, I think there is an opportunity to improve functionality by taking advantage
      of some of the other search categories (e.g., name, description, content type, units). This would
      be doubly useful since content type and units are not searched when performing a generic search.
    - [suggestion] Rather than automatically inserting wildcards (`*`), I would recommend informing the user of the wildcard
      character so that they can have finer control over how wildcards are used in their search.
    - It would be nice to mention in the documentation that several terms can be searched at once using
      a comma. Alternatively, the parameter could accept a character vector that could be
      collapsed with commas.
    - [suggestion] The function could potentially accept character vectors from each of the categories (as well as
      a generic search term field) and then the appropriate search string could be created.
- `nomis_codes()`
    - Generally, it is helpful to mirror the language of the source material when possible.
      The `nomis_codes()` function accepts a dimension (concept) and the function returns codes for that dimension.
      Therefore, the argument `code` might be renamed to `dimension` or `concept`, since the
      function searches for codes availalbe within a dimension.
    - Related to above, I think the `@title` could be more descriptive.
    - [suggestion] The query is currently built using a hierarchal structure. *E.g.*, `/api/v01/dataset/<ID>/<DIMENSION>/<CODE>/def.sdmx.json`. However,
      I think there is an opportunity to make this query a little more flexible and extensible. Instead of enforcing the hierarchy,
      the funciton could use the GET verb to provide a little more control over how the data is filtered. A user could
      query on criteria from different dimensions than the one which they are interested in results from. For example,
      even if the user is interested in the 'item' dimension, they may want to filter on 'geography' and 'sex'. Using GET would allow
      for this. As an example: `/api/v01/dataset/NM_1_1/item.def.sdmx.json?geography=2038432081&sex=7`. This would also expose the user
      to options for specifying times and dates, which cannot otherwise be done. *I fully acknowledge that this might be over-building,
      since this function only provides a support role and will probably not be used that way.*
    - This is a minor concern, but I feel it's cleaner to specify value if a test function returns TRUE first when using `ifelse()`.
      The code `ifelse(is.null(type), "", paste0("/", type))` reads cleaner than `ifelse(is.null(type) == FALSE, paste0("/", type), "")` IMO.
- `nomis_data_info()`
    - The documentation could be more explicit in telling the user that the data returned by this function is metadata.
    - [suggestion] Relating to the comment above, it may be a little clearer to rename the function something to the effect of `nomis_get_metadata()`.
      This naming scheme is more in line with the pattern set forth by `nomis_get_data` naming scheme.
- `nomis_get_data()`
    - I see that the API documentation states, "you have a 25000 **cell** limit on the number of data values downloaded in these formats."
      It appears that the code currently breaks the fetch into multiple steps if the record count is greater than or equal to 25k **records**.
      I am curious if this might be an oversight. It might be useful to test the edge case where the query returns more than 25k cells but
      fewer than 25k records to see if everything still works.
    - Including a `select` parameter may be a good enhancement for the future. Right now, that can be accomplished with
      `additional_queries` but is probably important enough to be standalone. This would allow for additional documentation
      and sugar to be added for it.

#### Conclusion

This is a great package that will be valuable to anyone interested in analysis of this data. Nice work!

I appreciate the opportunity to review this package and hope that my comments are constructive and useful. This package needs just a few
tweaks (namely getting the examples up and running) and then I will gladly approve!
