+++ R/data_download.R
@@ -1,234 +1,259 @@
 
 
 #' Retrieve Nomis datasets
-#' 
-#' Retrieves specific datasets from nomis, based on their ID. To find dataset 
-#' IDs, use \code{\link{nomis_data_info}}. Datasets are retrived in csv format 
-#' and parsed with the \code{read_csv} function from the \code{readr} package 
-#' into a tibble, with all columns parsed as character columns.
-#' 
-#' To find the code options for a given dataset, use \code{\link{nomis_codes}}.
-#' 
-#' This can be a slow process if querying significant amounts of data.
-#' 
-#' Note the difference between the \code{time} and \code{date} parameters. 
-#' The \code{time} and \code{date} parameters should not be used at the same 
-#' time. If they are, the function will retrieve data based on the the 
-#' \code{date} parameter. If given more than one query, \code{time} will 
-#' return all data available between those queries, inclusively, while 
-#' \code{date} will only return data for the exact queries specified. So 
-#' \code{time=c("first", "latest")} will return all data, while 
-#' \code{date=c("first", "latest")} will return only the first and latest 
+#'
+#' @description Retrieves specific datasets from nomis, based on their ID. To 
+#' find dataset IDs, use \code{\link{nomis_data_info}}. Datasets are retrived 
+#' in csv format and parsed with the \code{read_csv} function from the 
+#' \code{readr} package into a tibble, with all columns parsed as character 
+#' columns.
+#'
+#' @description To find the code options for a given dataset, use 
+#' \code{\link{nomis_get_metadata}}.
+#'
+#' @description This can be a slow process if querying significant amounts of 
+#' data. Guest users are limited to 25,000 rows per query, although 
+#' \code{nomisr} identifies queries that will return more than 25,000 rows, 
+#' sending individual queries and combining the results of those queries into 
+#' a single tibble.
+#'
+#' @description Note the difference between the \code{time} and \code{date} 
+#' parameters.
+#' The \code{time} and \code{date} parameters should not be used at the same
+#' time. If they are, the function will retrieve data based on the the
+#' \code{date} parameter. If given more than one query, \code{time} will
+#' return all data available between those queries, inclusively, while
+#' \code{date} will only return data for the exact queries specified. So
+#' \code{time=c("first", "latest")} will return all data, while
+#' \code{date=c("first", "latest")} will return only the first and latest
 #' data published.
-#' 
+#'
 #'
 #' @param id The ID of the dataset to retrieve.
-#' 
-#' @param time Parameter for selecting dates and date ranges. There are two 
-#' styles of values that can be used to query time. 
-#' 
-#' The first is one or more of \code{"latest"} (returns the latest available 
-#' data), \code{"previous"} (the date prior to \code{"latest"}), 
-#' \code{"prevyear"} (the date one year prior to \code{"latest"}) or 
-#' \code{"first"} (the oldest available data for the dataset). 
-#' 
-#' The second style is to use or a specific date or multiple dates, in the 
-#' style of the time variable codelist, which can be found using the 
-#' \code{\link{nomis_codes}} function.
-#' 
-#' Values for the \code{time} and \code{date} parameters should not be used 
-#' at the same time. If they are, the function will retrieve data based 
-#' on the the \code{date} parameter.
-#' 
-#' Defaults to \code{NULL}.
-#' 
-#' @param date Parameter for selecting specific dates. There are two styles 
-#' of values that can be used to query time. 
-#' 
[…]
