+++ R/utils-collect.R
@@ -4,16 +4,12 @@
 
 
 nomis_collect_util <- function(query) {
-    
-    df <- readr::read_csv(
-      paste0(base_url, query), 
-      col_types = readr::cols(.default = readr::col_character())
-      )
-    
-    df
-    
+  
+  df <- readr::read_csv(
+    paste0(base_url, query),
+    col_types = readr::cols(.default = readr::col_character())
+  )
+
+  df
+  
 }
-
-
-
-
