### 2018-03-07 da398cb
added contributing.md
