### 2018-03-09 e954cd7
update tests so they don't run on cran, vignette tweaks
