+++ tests/testthat/test_search.R
@@ -2,20 +2,19 @@ library(nomisr)
 context("nomis_search")
 
 test_that("nomis_search return expected format", {
+  skip_on_cran()
+  
   y <- nomis_search(keywords = "Claimants")
-
   expect_length(y, 14)
   expect_type(y, "list")
   expect_true(tibble::is_tibble(y))
 
   x <- nomis_search("*seekers*")
-
   expect_length(x, 14)
   expect_type(x, "list")
   expect_true(tibble::is_tibble(x))
 
   z <- nomis_search("*Seekers*")
-
   expect_length(z, 14)
   expect_type(z, "list")
   expect_true(tibble::is_tibble(z))
