+++ tests/testthat/test_get_data.R
@@ -2,6 +2,8 @@ library(nomisr)
 context("nomis_get_data")
 
 test_that("nomis_get_data return expected format", {
+  skip_on_cran()
+  
   z <- nomis_get_data(
     id = "NM_1_1", time = "latest",
     measures = c(20100, 20201), sex = 5,
