+++ tests/testthat/test_data_info.R
@@ -3,6 +3,8 @@ context("nomis_data_info")
 
 
 test_that("nomis_data_info return expected format", {
+  skip_on_cran()
+  
   x <- nomis_data_info()
 
   expect_length(x, 14)
