+++ vignettes/introduction.R
@@ -4,28 +4,52 @@ x <- nomis_data_info()
 head(x)
 
 ## ---- echo=TRUE----------------------------------------------------------
+y <- nomis_data_info("NM_893_1")
+
+tibble::glimpse(y)
+
+## ---- echo=TRUE----------------------------------------------------------
 library(dplyr, warn.conflicts = F)
 
-y <- nomis_data_info("NM_893_1")
 y$annotations.annotation %>% class()
+
+y$annotations.annotation %>% length()
 
 y$annotations.annotation[[1]] %>% class()
 
 y %>% pull(annotations.annotation) %>% class()
 
-y %>% pull(annotations.annotation)%>% .[[1]] %>% class()
+y %>% pull(annotations.annotation) %>% .[[1]] %>% class()
 
-y %>% pull(annotations.annotation) %>% purrr::pluck(1) %>% class()
+y %>% pull(annotations.annotation) %>% purrr::pluck() %>% class()
 
-y %>% pull(annotations.annotation) %>% purrr::map(1) %>% class()
-
+## Unnesting list columns
 y %>% tidyr::unnest(annotations.annotation) %>% glimpse()
 
 ## ---- echo=TRUE----------------------------------------------------------
-a <- nomis_search('*claimants*')
+a <- nomis_search(name = '*jobseekers*', keywords = 'Claimants')
+
+tibble::glimpse(a)
+
+a %>% tidyr::unnest(components.attribute) %>% glimpse()
+
+b <- nomis_search(keywords = c('Claimants', '*Year*'))
+
+tibble::glimpse(b)
+
+b %>% tidyr::unnest(components.attribute) %>% glimpse()
+
 
 ## ---- echo=TRUE----------------------------------------------------------
-b <- nomis_search(keywords = 'Claimants')
+q <- nomis_overview("NM_1650_1")
+
+q %>% tidyr::unnest(name) %>% glimpse()
+
+
+## ---- echo=TRUE----------------------------------------------------------
+ s <- nomis_overview("NM_1650_1", select = c("units", "keywords"))
+ 
+ s %>% tidyr::unnest(name) %>% glimpse()
 
 ## ---- echo=TRUE----------------------------------------------------------
 
@@ -51,10 +75,10 @@ b <- nomis_search(keywords = 'Claimants')
 ## ---- echo=TRUE----------------------------------------------------------
  z <- nomis_get_data(id = "NM_893_1", time = "latest", geography = "TYPE266")
 
- tibble::glimpse(z)
+print(z)
 
 ## ---- echo=TRUE----------------------------------------------------------
  x <- nomis_get_data(id = "NM_893_1", time = "latest", geography = c("1929380119", "1929380120"))
  
-tibble::glimpse(x)
+print(x)
 
