### 2018-03-09 #190 @evanodell
Hi @cderv and @pegeler, thank you both of you very much for your helpful comments and suggestions, apologies for not being able to respond to them sooner.

## Response to @cderv

> #### List-columns
>
> I find it kind of necessary because you also used list-columns in tibbles. This is not something easy for some newcomers to handles. Even for some dplyr users who do not used that a lot, it is not something easy. Adding some example in the vignettes to show how to retrieve embedded data.frame inside some tibble's list column could be a good idea. To illustrate:
>

```
> library(nomisr)
> library(dplyr, warn.conflicts = F)
> y <- nomis_data_info("NM_893_1")
> y$annotations.annotation %>% class()
> #> [1] "list"
> y$annotations.annotation[[1]] %>% class()
> #> [1] "data.frame"
> y %>% pull(annotations.annotation) %>% class()
> #> [1] "list"
> y %>% pull(annotations.annotation)%>% .[[1]] %>% class()
> #> [1] "data.frame"
> y %>% pull(annotations.annotation) %>% purrr::pluck(1) %>% class()
> #> [1] "data.frame"
> y %>% pull(annotations.annotation) %>% purrr::map(1) %>% class()
> #> [1] "list"
> y %>% tidyr::unnest(annotations.annotation) %>% glimpse()
> #> Observations: 14
> #> Variables: 11
> #> $ agencyid                             <chr> "NOMIS", "NOMIS", "NOMIS"...
> #> $ id                                   <chr> "NM_893_1", "NM_893_1", "...
> #> $ uri                                  <chr> "Nm-893d1", "Nm-893d1", "...
> #> $ version                              <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1...
> #> $ components.primarymeasure.conceptref <chr> "OBS_VALUE", "OBS_VALUE",...
> #> $ components.timedimension.codelist    <chr> "CL_893_1_TIME", "CL_893_...
> #> $ components.timedimension.conceptref  <chr> "TIME", "TIME", "TIME", "...
> #> $ name.value                           <chr> "LC4408EW - Tenure by num...
> #> $ name.lang                            <chr> "en", "en", "en", "en", "...
> #> $ annotationtext                       <chr> "Current (being actively ...
> #> $ annotationtitle                      <chr> "Status", "Keywords", "Un...
>

```
**Response:** I've added a similar example to the vignette, which is similar enough to your example that I feel I should cite you as a contributor if that is okay.

> Following are my review comments, for improvement:
>
> #### About Community Guidelines
>
> There is no contribution guidelines in the README or other CONTRIBUTING.md.

**Response:** Added a CONTRIBUTING.md in https://github.com/evanodell/nomisr/commit/e4bbb0229709a861b1ef2141e9ecf6da6ff8fc10

> DESCRIPTION needs some improvement : no need to use Author and Maintainer fields if you are using Authors@R. You can just use the latest. You can find some good example in R package book.

**Response:** Now only using Authors@R

> #### Improvement for the packaging guidelines
>
> see below parts for details
>
> * Adding a code of conduct

**Response:** I a code of conduct in https://github.com/evanodell/nomisr/commit/ede35580b8369aba016b8375f0d6569b58063e9c and

> * Improve README considering it is the first entry point for discovering NOMIS API
> * Improve DESCRIPTION for authorship
> * [minor] follow versionning advice with *.*.*.9000 for in development version

**Response:** Now following versioning best practice in development version
> * [question] I did not find why you need curl package in suggest...
**Response:** For reasons I'm not entirely clear on, the package was failing on Travis without the curl suggestion, I think due to a bug in jsonlite.

> #### README improvement
>
> * I would be more precise on what the NOMIS database is and for who it is. For example, I found that this submission issue mentioned UK database but not the README. You help file for the package as pretty clear description of what it is for. While reviewing, I had a doubt and went to check on NOMIS. This is why I think the statement of need is not complete.
> * It feels like the README could get more examples on what your package can do. Only one of your functions is shown in the readme.
> * Put package documentation link in the title of the repo like in the reprex repo. By the way, you can add a more descriptive description on your github repo and some tags.
> * could precise that it uses the tibble format (from the tidyverse)

**Response:** I've added more details to the README, including more details on Nomis itself, a simple example and citation information.

>
> #### Vignettes improvement
>
> * Vignette was not built when I installed the package with devtools::install_github. Did not know if this intended behavior or not.
> * I tried to build the vignette locally then: it worked.
>   * However, there is a problem on result formatting with the tibble. (I think because you use result = 'asis' in code chunks) and because every results are not tables (y$annotations.annotation is a list for example.) so it does not print nicely.
>     * The vignette do not show any code to get the results - I would show them for someone to learn how to use your function, as examples. (with echo = TRUE)
>     * The formatting issue is visible on your website too
> * I think the vignette contains necessary information but does not show enough examples on how to retrieve data in the objects you return. You mentioned the three list columns with nested data.frame but not show how correctly extracts the info.
> * It is pretty clear that you package allows to retrieve a large amount of data from the NOMIS API. it seems necessary to always use nomis_data_info, then nomis_codes then nomis_get_data to be sure to not retrieve to much. The vignette helps you understand that. It could appears to in your readme, and it could be more clear how to construct a pipeline for good practice with this API.
>
> These improvement are the reason why I did not check the vignettes box, even if there is one in the github repo and doc.
> If you need help on all this, do not hesitate to ask.

**Response:** As mentioned above, I've added your suggestions on exploring list-columns to the vignette, and fixed the online formatting issues. I've changed echo to TRUE to show code, pre built vignette in https://github.com/evanodell/nomisr/commit/1c9cd3f37d35f65b64f354d42a51c490cd0a6b36

> #### About documentation of functions and examples
>
> * All functions are well documented.
> * I encountered some issues running the examples because of a timeout. I believe this is what you meant in your by "there are some limits to the amount of data that can be retrieved within a certain period of time, although those are not published.". The maybe some improvement to do on this. (see under)
> * Is this the reason why you enclosed each example in \dontrun{} statement ? am I not sure on what is the recommendation for API package like this one for example ? \dontrun{} does allow anyone to run examples with something like example("nomis_codes", package = "nomisr"). Idea : Maybe \donttest{} is better here: does show in help and runs with example, but not run by R CMD check (so not run by CRAN ?) - see rd vignette from roxygen2

**Response:** I've changed \dontrun{} to \donttest{} in https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4

> * I ran all tests with devtools::run_examples(run = FALSE) (in several time, due to timeout constraint). Everything is working. It works well for almost all, but your example assign systematically to a variable (named x or y - room for improvement) but never print to show head of results. I feel that is missing.

**Response:** I have changed the names in examples in https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4 to be more descriptive of the data they are actually retrieving, and added a print call so the head is actually displayed.

> * I said "almost all" because there are some very long request in nomis_get_data. The warning that you put in Details states that it can be very slow process if calling significant amount of data. I would emphasize on this, saying this warning in Description part of the help file. And these example should definitely not be ran.

**Response:** I've removed the very long examples from nomis_get_data https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4 - I considered wrapping them in \dontrun{} but I think dropping them is the better option.

> * I saw you made a pkgdown website. This is great !! and nice theme!

**Response:** Thank you!

> *I did not find the source code of this website. I would have thought it leaved with the package source code. But no docs folder or gh-pages. I find it useful to have access for letting people PR your some typos or other thing and all the site from your package source code. However, you may have good reason.

**Response:** I keep all the source code for my documentation website in my [docs repo](https://github.com/evanodell/docs). I did this so I could host on netlify with a custom domain and https encryption and not have to use separate subdomains for each package. I'm not sure if that is the best option or not.

> ## Various comments on the code
>
> * [minor] I would put the url of the API in its own variable. I think it is better practice and more efficient to maintain in case of change in the url (e.g V0.2). The url is used several time and a change would mean copy pasting. another utils function could help building the url (paste0 is used three times), with the same base url def.sdmx.json). Know that it exists some tools to help build some strings with parameters like glue, and specific one to handle url like urltools or those in httr. You may not need them here but it is worth knowing.

**Response:** I've turned the API URL into its own variable in https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4

> * suppressMessage is used to hide the message of readr::read_csv. However, the message appears because column types are guessed and not set. It is good practice to set column type in readr with column_types = cols(<something>). In an API package like this one, it make even more sense because it could helps be aware of any change in the format. (using stop_for_problems() or just see a message appear) - see readr vignettes

**Response:** I've dropped suppressMessage and used `col_types = readr::cols(.default = readr::col_character())` to make columns characters by default in https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4

>
> * [minor] I don't know what are the good practice on this one. I just feel that you have a dplyr dependency in import for three functions bind_rows, case_when and if_else than can be easily replace with rbind or if.else. I feel that, if the former are very useful interactively in an analysis, using them in a package can be unnecessary dependency to a "big" package that can cause you some effort later on if anything change in dplyr. More of a personal point of view on this one.
>
> However, it relates on how you are build your query url
>
> ```
>   query <- dplyr::if_else(keywords==FALSE,
>                            paste0("/def.sdmx.json?search=*",
>                                   search, "*"),
>                            paste0("/def.sdmx.json?search=keywords-",
>                                   search, "*")
> ```
>
> could be rewritten without dplyr::if_else
>
> ```
> base_query <- "/def.sdmx.json?search="
> if (keyword) {
>     keyword_term <- "keywords-")
> } else {
>     keyword_term <- "*"
> }
> # or keyword_term <- if.else(keyword, "keywords-", "*")
> query <- paste0(base_query, key_word_term, search, "*")
> ```
>
> It would suppres a non essential dependency.
>

**Response:** I've dropped the use of case_when and if_else in favour the the style you described above in https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4.  `bind_rows` is used to combine together a single tibble with a list of one or more tibbles, and using rbind in place of bind_rows would likely require recreating the `bind_rows` function. If there are changes to `bind_rows` in the future it is only in one location so should be easy to adjust to.

>
> * The style in code could be improve in some places. You can follow some style guide to help you. like the one in ROpensci Guidelines linking to Advanced R or more recent the one used in the tidyverse styleguide. They even build some tools to help like styler and lintr. Example of improvement:
>        * put spaces after foror if and also around =, ==, +, >, ...
>        * indentation of code (in RStudio CTRL+i i useful). for example in if () { } else { }, lines of code that are inside brackets should be indented.
>         * use {} with if when it does not fit on one line
>        * More importantly, be consistent in all your script with the guideline you choose.
> * Know that when you want to stop on a condition, there is also stopifnot. For user experience, you can also return clearer error sometimes with stop("API request did not return any results", call. = FALSE) to not show the call for example.
> * A good practice is also to use meaningful name inside your internal code, even for temporary variable. There are a lot of x, a, q, ... This will improve readability and the ease for others to collaborate on your code (including future you)

**Response:** I've added `call. = FALSE` to the various stop messages.

**Response:** I've gone through the code and changed temporary variable names to be more meaningful, and formatted the code to improve indentation, spaces around if and for operators, ==, etc.

> ## About the timeout and size of data
>
> *If session is interactive(), it may be a good idea to warn clearly and even ask if we want to continue when lots of data are being requested. I find myself playing with the example trying to retrieve 208888 of your pages
>
> ```
> Retrieving additional pages 1 of 208888
> Retrieving additional pages 2 of 208888
> ```
>
> I don't know if some other API packages are doing something to help user or if it is all the user responsibility to know what is he doing. (users will surely be already NOMIS data users).
>
>     As another improvement, I think you should try to create custom error message for the timeout issue as you know it is an issue that will impact user experience but cannot really prevent. I encountered:
>
> ```
> a <- nomis_codes("NM_1_1")
> # Error in open.connection(con, "rb") :
> #  Timeout was reached: Resolving timed out after 10000 milliseconds
> ```
>
> You could use some tryCatch mechanism (some useful info in Advanced R).
>
> * [thoughts] You may need more control on the request mechanism to deal with timeout. First, it seems for the website one need to contact NOMIS help-desk to know more about limitation. Then , your are relying on readr connection mechanism, and may need to use httr, crul or curl to better handle request to API with the specificity of NOMIS
>

**Response:** I've used `tryCatch` to provide more useful information in the case of time outs. I've also written, but not implemented, a warning and yes/no option in interactive sesssions with requests over 350,000 cells which are likely to cause rate limiting issues. Implemented in https://github.com/evanodell/nomisr/commit/ac3e5a4fc92836e8736d5c47e9312838dd0628e4

> ## About SMDX format on NOMIS
>
>     Do you know about the rsdmx package ? it is on github and CRAN, there is WIKI. I did not find NOMIS in their listing, so I did not know if it works for you and how it could overlap with your package.
>     rsdmx is definitely more a generic package than yours. I just wanted to point it to you in case you did not know.

**Response:** I was not aware of it, but it is something I should explore for the future, thank you for pointing me towards it.

 ## Response to pegeler

> Package Review for nomisr
>
> Onboarding submission: ropensci/onboarding#190
>
> Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide
>
>     As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor before starting your review).
>
> Documentation
>
> The package includes all the following forms of documentation:
>
>     A statement of need clearly stating problems the software is designed to solve and its target audience in README
>     Installation instructions: for the development version of package and any non-standard dependencies in README
>     Vignette(s) demonstrating major functionality that runs successfully locally
>     Function Documentation: for all exported functions in R help
>     Examples for all exported functions in R Help that run successfully locally
>     Community guidelines including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).
>
>     For packages co-submitting to JOSS
>
>         The package has an obvious research application according to JOSS's definition
>
>     The package contains a paper.md matching JOSS's requirements with:
>
>         A short summary describing the high-level functionality of the software
>         Authors: A list of authors with their affiliations
>         A statement of need clearly stating problems the software is designed to solve and its target audience.
>         References: with DOIs for all those that have one (e.g. papers, datasets, software).
>
> Functionality
>
>     Installation: Installation succeeds as documented.
>     Functionality: Any functional claims of the software been confirmed.
>     Performance: Any performance claims of the software been confirmed.
>     Automated tests: Unit tests cover essential functions of the package
>     and a reasonable range of inputs and conditions. All tests pass on the local machine.
>     Packaging guidelines: The package conforms to the rOpenSci packaging guidelines
>
> Final approval (post-review)
>
>     The author has responded to my review and made changes to my satisfaction. I recommend approving this package.
>
> Estimated hours spent reviewing: 10
> Review Comments
>
> Excellent work! This is a fine example of the level of quality and professionalism that I have
> come to expect from rOpenSci contributions. This package is a great addition to the
> complement of R functionality and provides a facile portal to a useful data source.
>
> Many of the comments that I had have been addressed as a result of @cderv's prompt and thorough review.
> In addition, development has continued to be very active over the past few weeks! Issues are
> fixed as quickly as I find them! Below are a few more comments that I would like to add.
> Basic Package Structure Overview
> README.md

The README.md does not contain citation information, per rOpenSci packaging guide.
I am not sure how necessary that is since there is a CITATION file in the inst/ directory,
however, it should be an easy add.

**Reponse:** Added in https://github.com/evanodell/nomisr/commit/b75a47dd503c47337ecccf2691ef3c48573154fc

> I would concur with @cderv in suggesting additional usage examples.
> DESCRIPTION
>
>     The DESCRIPTION file does not currently specify a minimum version of R. I would
>     recommend doing so. A minor release is preferred over a patch release, e.g.
>
>     Depends: R (>= 3.4.0)
>
>     is better than
>
>     Depends: R (>= 3.4.1)
>
>     I am not seeing where curl is used even though it is in the Suggests list.

**Response:** I've added a minimum version of R (>= 3.4.0) to the DESCRIPTION in https://github.com/evanodell/nomisr/commit/b75a47dd503c47337ecccf2691ef3c48573154fc

**Response:** Not including curl was causing errors with jsonlite on travis - I'm not sure why, as jsonlite is supposed to import curl itself.

> ## CITATION

[…]
