### 2018-03-10 #190 @cderv
Hi @evanodell

Great work ! And thanks a lot for your answer formatting just above. It is very easy to follow how you took our comments into account. Nice.

I updated my first post by checking some boxes ✅  and ⏲ of review.

I still have a few remarks after taking another pass on your updated code

### Package Website

Nice idea to centralize all your documentation. However, there is still a few things to adjust as I get 404 error navigating through doc. I can only view the reference page of the package but not the vignette for example. Must be something to do with how `pkgdown` deals with path when hosting in another site.
I do not have the solution for your https request, but you may find some help in the #rstats community.

###  Contributing.md

Great addition. Two things:

+ I would add it to `.Rbuildignore` as you get a R CMD CHECK note otherwise
+ I would add a link to this doc in the README. ([Example in another great Ropensci 📦 ](https://github.com/ropensci/drake#contributing))

### Readme

Great improvement !! I find it now a good introduction to your 📦 with nice examples.
You may find me sticky on this but the mention to UK is still missing in your introduction. Your `DESCRIPTION`  is good, so why not add the first line in your readme ?

> Access UK official statistics from the Nomis database through R

About formatting, take care : links won't works when between single quote \`\`.
```
`[tibbles](http://tibble.tidyverse.org/)`
```
must be changed to
```
[`tibbles`](http://tibble.tidyverse.org/)
```
in order to get [`tibbles`](http://tibble.tidyverse.org/)

Also, you need to keep in mind to synchronise your `README.Rmd` modification and `README.md` generation. If you use `usethis::use_readme_rmd`, it puts a git pre-commit hook that prevent you to commit if `md` is older than `Rmd`. That way you'll never forget to rerender. This comes from `devtools::use_readme_rmd` now in `usethis`. (BTW, a _pre-commit hook_ is a little `.sh` script that is executed before every commit. You can see it [there](https://github.com/r-lib/usethis/blob/master/inst/templates/readme-rmd-pre-commit.sh))

### Vignettes

It looks good !  No problem to take the example I gave you. It was just to illustrate, it could be rework but if you found it useful to you, it could be useful to other.
The html vignette is now included, but not up to date. Before installing the package, after a `git pull` I needed to do a `devtools::build_vignette`. Not sure if the vignette are to be built in source package. I found re-reading [R package book]() that they can be built at installation with `devtools::install_github("evanodell/nomisr", build_vignettes = TRUE)`, so i guess it is not an issue if not present in the package.

### about curl and travis

I read something on that but I did not found it. Sorry. It may worth asking on ropencsi [forum](https://discuss.ropensci.org/), rstudio [community](community.rstudio.com).

### About download data limit

I saw you wanted to do something with `interactive()` in your code, but it is still commented out. You may try to put this code in separate feature branch to clean your master if it is still work in progress.

I think your 📦 is ready ! Very nice work, I still find it a simple one but powerful one as always with data packages. Thanks!
