+++ README.Rmd
@@ -25,7 +25,7 @@ knitr::opts_chunk$set(
 
 `nomisr` is for accessing [Nomis](https://www.nomisweb.co.uk/) data with R. Nomis contains data from the Census, the Labour Force Survey, DWP benefit statistics and other economic and demographic data, and is maintained on behalf of the Office for National Statistics by the University of Durham.
 
-The `nomisr` package provides functions to find what data is available, the  variables and query options for different datasets and a function for downloading data. `nomisr` returns data in [tibble](https://cran.r-project.org/package=tibble) format. Most of the data available through `nomisr` is based around statistical geographies, with a handful of exceptions.
+The `nomisr` package provides functions to find what data is available, the  variables and query options for different datasets and a function for downloading data. `nomisr` returns data in [`tibble`](https://cran.r-project.org/package=tibble) format. Most of the data available through `nomisr` is based around statistical geographies, with a handful of exceptions.
 
 The package is for demographers, economists, geographers, public health researchers and any other researchers who are interested in geographic factors. The package aims to aid reproducibility, reduce the need to manually download area profiles, and allow easy linking of different datasets covering the same geographic area.
 
@@ -40,7 +40,7 @@ devtools::install_github("evanodell/nomisr")
 
 ## Using `nomisr`
 
-`nomisr` contains functions to search for datasets, identify the query options for different datasets and retrieve data from queries, all done with `[tibbles](http://tibble.tidyverse.org/)`, to take advantage of how `tibble` manages list-columns. The use of metadata queries, rather than simply downloading all available data, is useful to avoid overwhelming the rate limits of the API. For full details on all available functions and demonstrations of their use, please see the package [vignette](https://docs.evanodell.com/nomisr/articles/introduction.html).
+`nomisr` contains functions to search for datasets, identify the query options for different datasets and retrieve data from queries, all done with [`tibbles`](http://tibble.tidyverse.org/), to take advantage of how `tibble` manages list-columns. The use of metadata queries, rather than simply downloading all available data, is useful to avoid overwhelming the rate limits of the API. For full details on all available functions and demonstrations of their use, please see the package [vignette](https://docs.evanodell.com/nomisr/articles/introduction.html).
 
 The example below gets the latest data on Jobseeker's Allowance with rates and proportions, on a national level, with all male claimants and workforce.
 
@@ -97,5 +97,4 @@ A BibTeX entry for LaTeX users is
 ```
 
 License: MIT
-
 
