+++ README.Rmd
@@ -98,5 +98,5 @@ A BibTeX entry for LaTeX users is
   }
 ```
 
-License: MIT
+License: [MIT](LICENSE.md)
 
