### 2018-03-14 d116098
citation cleanup
