+++ README.Rmd
@@ -81,9 +81,7 @@ Please note that this project is not affiliated with the Office for National Sta
 
 Get citation information for `nomisr` in R with `citation(package = 'nomisr')`
 
-Odell E (2018). _nomisr: Access Nomis UK labour market data on with R_. doi: 10.5281/zenodo.1157908 (URL:
-http://doi.org/10.5281/zenodo.1157908), R package version 0.0.2.9000, <URL:
-https://docs.evanodell.com/nomisr>.
+Odell E (2018). _nomisr: Access Nomis UK labour market data on with R_. doi: 10.5281/zenodo.1157908, R package version 0.0.2.9000, <URL: https://docs.evanodell.com/nomisr>.
 
 A BibTeX entry for LaTeX users is
 
