+++ DESCRIPTION
@@ -24,7 +24,6 @@ Suggests:
   knitr,
   rmarkdown,
   testthat,
-  curl,
   purrr,
   tidyr
 VignetteBuilder: knitr
