+++ man/nomis_api_key.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/api-key.R
+\name{nomis_api_key}
+\alias{nomis_api_key}
+\title{API Key}
+\usage{
+nomis_api_key(force = FALSE)
+}
+\arguments{
+\item{force}{}
+}
+\description{
+API Key
+}
