+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(nomis_api_key)
 export(nomis_content_type)
 export(nomis_data_info)
 export(nomis_get_data)
