+++ DESCRIPTION
@@ -24,6 +24,7 @@ Suggests:
   knitr,
   rmarkdown,
   testthat,
+  curl,
   purrr,
   tidyr
 VignetteBuilder: knitr
