+++ R/api-key.R
@@ -9,11 +9,12 @@ nomis_api_key <- function(force = FALSE) {
   if (!identical(env, "") && !force) return(env)
   
   if (!interactive()) {
-    stop("Please set env var NOMIS_API_KEY to your Nomis API key",
+    stop("Please set environment variable NOMIS_API_KEY to your Nomis API key",
          call. = FALSE)
   }
   
-  message("Couldn't find env var NOMIS_API_KEY See ?nomis_api_key for more details.")
+  message("Couldn't find environment variable NOMIS_API_KEY") 
+  message("See ?nomis_api_key for more details.")
   message("Please enter your API key and press enter:")
   key <- readline(": ")
   
