### 2018-03-19 1fb95ef
trying out some things with travis and curl
