### 2018-03-22 d32dcba
added more tests to cover API key
