+++ R/api-key.R
@@ -1,6 +1,16 @@
 
-#' API Key
-#' @param force 
+#' Nomis API Key
+#' 
+#' @description The Nomis API has an optional key. Using the key means that 
+#' 100,000 rows can be returned per call, which can speed up larger data 
+#' requests and reduce the chances of being rate limited or having requests 
+#' timing out.
+#' 
+#' @description You can sign up for an API key 
+#' \href{https://www.nomisweb.co.uk/myaccount/userjoin.asp}{here}.
+#' 
+#' @param force If TRUE, resets the API key and requires a 
+#' new key to be provided.
 #'
 #' @export
 nomis_api_key <- function(force = FALSE) {
@@ -11,14 +21,15 @@ nomis_api_key <- function(force = FALSE) {
   if (!interactive()) {
     stop("Please set environment variable NOMIS_API_KEY to your Nomis API key",
          call. = FALSE)
+    message("Use `Sys.setenv(NOMIS_API_KEY = <key>)`")
   }
   
   message("Couldn't find environment variable NOMIS_API_KEY") 
-  message("See ?nomis_api_key for more details.")
   message("Please enter your API key and press enter:")
   key <- readline(": ")
   
   if (identical(key, "")) {
+    Sys.unsetenv('NOMIS_API_KEY')
     stop("Nomis API key entry failed", call. = FALSE)
   }
   
