+++ .travis.yml
@@ -2,6 +2,7 @@
 
 language: r
 r:
+  - release
   - devel
 
 sudo: required
