+++ .travis.yml
@@ -3,6 +3,7 @@
 language: r
 r:
   - release
+  - devel
 
 sudo: required
 cache: packages
