### 2018-03-22 a1f8563
drop devtools and curl from travis
