### 2018-03-22 7c87932
update inclusion of loadpkg
