+++ NAMESPACE
@@ -8,6 +8,7 @@ export(nomis_overview)
 export(nomis_search)
 importFrom(dplyr,bind_rows)
 importFrom(jsonlite,fromJSON)
+importFrom(jsonlite,loadpkg)
 importFrom(readr,read_csv)
 importFrom(tibble,as_tibble)
 importFrom(tibble,enframe)
