### 2018-03-22 e79f110
another jsonlite fix attempt
