+++ NAMESPACE
@@ -6,9 +6,8 @@ export(nomis_get_data)
 export(nomis_get_metadata)
 export(nomis_overview)
 export(nomis_search)
+import(jsonlite)
 importFrom(dplyr,bind_rows)
-importFrom(jsonlite,fromJSON)
-importFrom(jsonlite,loadpkg)
 importFrom(readr,read_csv)
 importFrom(tibble,as_tibble)
 importFrom(tibble,enframe)
