+++ R/utils-get-data.R
@@ -7,10 +7,27 @@ nomis_get_data_util <- function(query) {
   ### Implement tryCatch
   # show_condition <- function(code) {
   #   tryCatch(
+  
+  api_get <- httr::GET(paste0(base_url, query))
+  
+  if (httr::http_type(api_get) != "text/csv") {
+    stop("Nomis API did not return data in required CSV format", call. = FALSE)
+  }
+  
+  if (httr::http_error(api_get)) {
+    stop(
+      sprintf(
+        "Nomis API request failed with status ", 
+        httr::status_code(api_get)
+      ),
+      call. = FALSE
+    )
+  }
+  
   df <- tryCatch(
     {
       readr::read_csv(
-        paste0(base_url, query),
+        api_get$url,
         col_types = readr::cols(.default = readr::col_character())
       )
     }, error = function(cond) {
