+++ R/utils-query.R
@@ -3,7 +3,23 @@
 
 
 nomis_query_util <- function(query) {
-  avail_response <- jsonlite::fromJSON(paste0(base_url, query), flatten = TRUE)
+  
+  api_resp <- httr::GET(paste0(base_url, query))
+  if (http_type(api_resp) != "application/json") {
+    stop("Nomis API did not return data in required json format", call. = FALSE)
+  }
+  
+  if (httr::http_error(api_get)) {
+    stop(
+      sprintf(
+        "Nomis API request failed with status ", 
+        httr::status_code(api_get)
+      ),
+      call. = FALSE
+    )
+  }
+  
+  avail_response <- jsonlite::fromJSON(api_resp$url, flatten = TRUE)
 
   df <- tibble::as_tibble(avail_response$structure$keyfamilies$keyfamily)
 
