+++ NAMESPACE
@@ -8,6 +8,10 @@ export(nomis_overview)
 export(nomis_search)
 import(jsonlite)
 importFrom(dplyr,bind_rows)
+importFrom(httr,GET)
+importFrom(httr,http_error)
+importFrom(httr,http_type)
+importFrom(httr,status_code)
 importFrom(readr,read_csv)
 importFrom(tibble,as_tibble)
 importFrom(tibble,enframe)
