### 2018-03-22 2c6badd
checking api response properly in utils-query
