### 2018-03-22 fedd306
added httr requirement
