+++ NAMESPACE
@@ -6,12 +6,12 @@ export(nomis_get_data)
 export(nomis_get_metadata)
 export(nomis_overview)
 export(nomis_search)
-import(jsonlite)
 importFrom(dplyr,bind_rows)
 importFrom(httr,GET)
 importFrom(httr,http_error)
 importFrom(httr,http_type)
 importFrom(httr,status_code)
+importFrom(jsonlite,fromJSON)
 importFrom(readr,read_csv)
 importFrom(tibble,as_tibble)
 importFrom(tibble,enframe)
