### 2018-03-22 d52b1d8
try dropping curl again
