+++ R/utils-get-data.R
@@ -16,9 +16,8 @@ nomis_get_data_util <- function(query) {
   
   if (httr::http_error(api_get)) {
     stop(
-      sprintf(
-        "Nomis API request failed with status ", 
-        httr::status_code(api_get)
+      paste0("Nomis API request failed with status ",
+                     httr::status_code(api_resp)
       ),
       call. = FALSE
     )
