### 2018-03-22 6cb7443
more testing for nomis_data_info
