### 2018-03-22 #190 @evanodell
Thank you @cderv

I did not know about `usethis::use_readme_rmd`, that has made things much easier. I've made the other tweaks you've suggested as well.

I've split off the the `interactive()` bits into a dev branch.

The curl and travis issue seems to stem from the internal `loadpkg` function in `jsonlite`. Using `httr` to check for http response codes seems to have solved the problem. I'm not sure why but at least the solution provides more informative error messages.

Regarding pkgdown, `path` in `build_site()` seems to have stopped working, but I think I now have a suitable workaround.
