### 2018-03-26 b34f17e
some weird failings with httr and nomis_data_info, this is just a hotfix
