### 2018-03-27 86e2c24
better error testing
