+++ tests/testthat/test_data_info.R
@@ -11,15 +11,18 @@ test_that("nomis_data_info return expected format", {
   expect_type(x, "list")
   expect_true(tibble::is_tibble(x))
 
-  expect_error(nomis_data_info("lalala"))
+  expect_error(nomis_data_info("lalala"), 
+               "Nomis API did not return data in required json format")
   
-  y <- nomis_data_info("NM_1_6")
-  expect_length(y, 14)
+  y <- nomis_data_info("NM_1658_1")
+  expect_length(y, 12)
   expect_type(y, "list")
   expect_true(tibble::is_tibble(y))
   expect_equal(nrow(y), 1)
-  expect_equal(y$id, "NM_6_1")
+  expect_equal(y$id, "NM_1658_1")
   expect_true(y$uri %in% x$uri)
   
-  
+  expect_error(nomis_data_info("NM_1_6"), 
+               "API request did not return any results")
+
 })
