### 2018-03-28 7d0a8db
Loading dplyr (dependency) for examples in nomis_overview().
