+++ R/overview.R
@@ -55,6 +55,8 @@
 #'
 #' @examples \donttest{
 #'
+#' library(dplyr)
+#' 
 #' q <- nomis_overview("NM_1650_1")
 #'
 #' q %>% tidyr::unnest(name) %>% glimpse()
