### 2018-03-28 9850811
Adding pre-built vignettes.
