+++ inst/doc/introduction.Rmd
@@ -1,0 +1,84 @@
+---
+title: "Introduction to nomisr"
+author: "Evan Odell"
+date: "`r Sys.Date()`"
+# output: rmarkdown::html_vignette
+output: 
+  html_document:
+    toc: true
+    toc_float: true
+    theme: flatly
+vignette: >
+  %\VignetteIndexEntry{Introduction to nomisr}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+# Introduction to the Nomis API
+
+Nomis provides free access to up-to-date official statistics from the Office for National Statistics (ONS), and is operated by the University of Durham.
+
+There is a lot of data available through Nomis, and there are some limits to the amount of data that can be retrieved within a certain period of time, although those are not published. For more details, see the [full API documentation](https://www.nomisweb.co.uk/api/v01/help) from Nomis.
+
+Nomis data is based around administrative and statistical geographies, and a particular geography should be specified when downloading data.
+
+# Accessing Nomis data
+
+`nomisr` is designed around a pipeline of three key functions: `nomis_data_info()`, `nomis_codes()` and `nomis_get_data()`.
+
+
+## Querying data availability
+
+Use the `nomis_data_info()` function without any parameters to get a tibble 
+with metadata for all available datasets:
+
+```{r}
+  library(nomisr) 
+  x <- nomis_data_info()
+  head(x)
+```
+
+
+`nomis_data_info()` can also be used to query metadata from a specific dataset, using its ID. The example below uses the "LC4408EW - Tenure by number of persons per bedroom in household by household type" dataset from the 2011 census, including the 3 nested data frames including metadata important to making specific queries.
+
+```{r}
+  y <- nomis_data_info("NM_893_1")
+  tibble::glimpse(y)
+  
+  y$annotations.annotation
+  
+  y$components.attribute
+  
+  y$components.dimension
+```
+
+
+When a tibble with metadata for all datasets or a specific dataset is returned, three of the columns: `annotations.annotation`, `components.attribute` and `components.dimension` are lists of nested data frames. `annotations.annotation` contains metadata on the dataset, including units and current status. `components.attribute` contains more detailed status metadata. `components.dimension` contains the grouping and summary variables available in the dataset, which vary between different datasets.
+
+## Downloading data
+
+Vast amounts of data are available through Nomis and so to avoid overwhelming the API, it is good practice to query what variables are available, using `nomis_codes`. This queries the different kinds of geography codes available through the API for the "LC4408EW - Tenure bynumber of persons per bedroom in household by household type" dataset.
+
+```{r}
+  a <- nomis_codes(id = "NM_893_1", code = "geography", type = "type")
+  a
+
+```
+
+
+Using the information above, we can now query the latest data on bedroom occupancy per household type in different NHS clinical commissioning groups.
+
+```{r}
+ z <- nomis_get_data(id = "NM_893_1", time = "latest", geography = "TYPE266")
+ tibble::glimpse(z)
+```
+
+
+
