+++ inst/doc/introduction.Rmd
@@ -2,83 +2,181 @@
 title: "Introduction to nomisr"
 author: "Evan Odell"
 date: "`r Sys.Date()`"
-# output: rmarkdown::html_vignette
 output: 
-  html_document:
-    toc: true
-    toc_float: true
-    theme: flatly
+    html_document:
+        toc: true
+        toc_float: true
+        theme: flatly
 vignette: >
   %\VignetteIndexEntry{Introduction to nomisr}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setup, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
 # Introduction to the Nomis API
 
-Nomis provides free access to up-to-date official statistics from the Office for National Statistics (ONS), and is operated by the University of Durham.
+`nomisr` is for accessing [Nomis](https://www.nomisweb.co.uk/) data with R. The Nomis API is free to access without registration, and contains up-to-date official statistics, including data from the Census, the Labour Force Survey, DWP benefit statistics and other economic and demographic data. Nomis is maintained on behalf of the Office for National Statistics by the University of Durham.
 
 There is a lot of data available through Nomis, and there are some limits to the amount of data that can be retrieved within a certain period of time, although those are not published. For more details, see the [full API documentation](https://www.nomisweb.co.uk/api/v01/help) from Nomis.
 
 Nomis data is based around administrative and statistical geographies, and a particular geography should be specified when downloading data.
 
-# Accessing Nomis data
-
-`nomisr` is designed around a pipeline of three key functions: `nomis_data_info()`, `nomis_codes()` and `nomis_get_data()`.
-
+`nomisr` is designed around a pipeline of three key functions: `nomis_data_info()`, `nomis_get_metadata()` and `nomis_get_data()`. The `nomis_overview()`, `nomis_content_type()` and `nomis_search()` functions can assist with this.
 
 ## Querying data availability
 
-Use the `nomis_data_info()` function without any parameters to get a tibble 
-with metadata for all available datasets:
+The `nomis_data_info()` function is focused on the structure and coverage of the available datasets.
 
-```{r}
-  library(nomisr) 
-  x <- nomis_data_info()
-  head(x)
+Use the `nomis_data_info()` function without any parameters to get a tibble with metadata for all available datasets:
+
+```{r, echo=TRUE}
+library(nomisr)
+x <- nomis_data_info()
+head(x)
 ```
 
 
-`nomis_data_info()` can also be used to query metadata from a specific dataset, using its ID. The example below uses the "LC4408EW - Tenure by number of persons per bedroom in household by household type" dataset from the 2011 census, including the 3 nested data frames including metadata important to making specific queries.
+`nomis_data_info()` can also be used to query metadata from a specific dataset, using its ID. The example below uses the "LC4408EW - Tenure by number of persons per bedroom in household by household type" dataset from the 2011 census, which has the ID "NM_893_1". 
 
-```{r}
-  y <- nomis_data_info("NM_893_1")
-  tibble::glimpse(y)
-  
-  y$annotations.annotation
-  
-  y$components.attribute
-  
-  y$components.dimension
+```{r, echo=TRUE}
+y <- nomis_data_info("NM_893_1")
+
+tibble::glimpse(y)
+```
+
[…]
