### 2018-03-30 ea93b97
A few minor edits prior to submitting second round of review.
