### 2018-03-30 #190 @pegeler
## Package Review for `nomisr`

Onboarding submission: ropensci/onboarding#190

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Overall this package is well put together.

The data structures returned by the Nomis API are very complex and this package does a good job of helping the user along to find the data they need. I was able to get metadata and data out without too much trouble. Existing Nomis users should find this package facilitates their workflow and users new to Nomis data will likely struggle a bit, but less so because of the package's enhanced vignettes, documentation, and README.

As development continues, I would suggest continuing to consider how to make the data more accessible to users who may not be familiar with data structures returned by the Nomis API. In doing so, the package user-base will continue to expand.

It appears that functionality is fairly complete. I am able to navigate metadata several ways and can pull data with relative ease. Great work!! **I think this package is ready!!!**

#### NEWS.md

The NEWS.md will need to be updated when 0.0.3 is released. Don't forget to `git tag` it as well 😀!

#### Conclusion

From the time of initial review to now, I have seen this package transformed from an already good package to something really awesome. I appreciate the way that the package author has taken reviewer comments and turned them into opportunities to rethink existing code. These changes represent a very thoughtful approach that is reflected in package usability and performance.

Thanks again for the chance to review!
