### 2018-03-31 #190 @cderv
@evanodell Well done! I am really impress with the evolution of this really a great 📦 from the beginning of the review and now. Awesome work ! Thanks for having taken into account our advices, I really appreciated it like @pegeler.

It was really a pleasure to review your package and it gave me the chance to discover the NOMIS database. I think your package is a great example of using a simple api and prepare the data return to an easier and practical format for the users.

I agree with @pegeler, this package is more than ready to ✨ in ropensci 📦 collections!

@karthik, what are the following steps ?
