### 2018-04-16 #190 @karthik
Congrats @evanodell! Your submission is now approved and accepted! 🎉 Thank you for submitting and many thanks to @cderv and @pegeler for thorough and timely reviews. To-dos:

```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)

```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
