+++ README.md
@@ -7,14 +7,14 @@
 MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/licenses/MIT)
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/nomisr)](https://cran.r-project.org/package=nomisr)
 [![GitHub
-tag](https://img.shields.io/github/tag/evanodell/nomisr.svg)](https://github.com/evanodell/nomisr)
+tag](https://img.shields.io/github/tag/ropensci/nomisr.svg)](https://github.com/ropensci/nomisr)
 [![](https://cranlogs.r-pkg.org/badges/grand-total/nomisr)](https://dgrtwo.shinyapps.io/cranview/)
 [![Travis-CI Build
-Status](https://travis-ci.org/evanodell/nomisr.svg?branch=master)](https://travis-ci.org/evanodell/nomisr)
+Status](https://travis-ci.org/ropensci/nomisr.svg?branch=master)](https://travis-ci.org/ropensci/nomisr)
 [![AppVeyor Build
-Status](https://ci.appveyor.com/api/projects/status/github/evanodell/nomisr?branch=master&svg=true)](https://ci.appveyor.com/project/evanodell/nomisr)
+Status](https://ci.appveyor.com/api/projects/status/github/ropensci/nomisr?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/nomisr)
 [![Coverage
-Status](https://img.shields.io/codecov/c/github/evanodell/nomisr/master.svg)](https://codecov.io/github/evanodell/nomisr?branch=master)
+Status](https://img.shields.io/codecov/c/github/ropensci/nomisr/master.svg)](https://codecov.io/github/ropensci/nomisr?branch=master)
 [![DOI](https://zenodo.org/badge/118144805.svg)](https://zenodo.org/badge/latestdoi/118144805)
 
 `nomisr` is for accessing [Nomis](https://www.nomisweb.co.uk/) data with
@@ -42,16 +42,18 @@ You can install `nomisr` from github with:
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("evanodell/nomisr")
+devtools::install_github("ropensci/nomisr")
 ```
 
 ## Using `nomisr`
 
 `nomisr` contains functions to search for datasets, identify the query
-options for different datasets and retrieve data from queries. The use
-of queries, rather than simply downloading all available data, is useful
-to avoid overwhelming the rate limits of the API. For full details,
-please see the package
+options for different datasets and retrieve data from queries, all done
+with [`tibbles`](http://tibble.tidyverse.org/), to take advantage of how
+`tibble` manages list-columns. The use of metadata queries, rather than
+simply downloading all available data, is useful to avoid overwhelming
+the rate limits of the API. For full details on all available functions
+and demonstrations of their use, please see the package
 [vignette](https://docs.evanodell.com/nomisr/articles/introduction.html).
 
 The example below gets the latest data on Jobseeker’s Allowance with
@@ -115,9 +117,9 @@ workforce.
  tibble::glimpse(z)
 #> Observations: 70
 #> Variables: 34
-#> $ DATE                <chr> "2018-01", "2018-01", "2018-01", "2018-01"...
-#> $ DATE_NAME           <chr> "January 2018", "January 2018", "January 2...
[…]
