### 2018-04-16 #190 @evanodell
Thanks @karthik! I'm more than happy to write something up on the package, whichever format you prefer.
