+++ cran-comments.md
@@ -1,0 +1,21 @@
+
+## Release summary
+
+This is the first submission of the `nomisr` package, with the version 0.1.0.
+`nomisr` provides functions to download UK census data and other official 
+statistics from the nomis API.
+
+## Test environments
+* local OS X install, R 3.4.4
+* ubuntu 14.04 (on travis-ci), R 3.4.4
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 2 note
+
+* This is a new release.
+
+* Possibly mis-spelled words in DESCRIPTION:
+  DWP (9:67)
+  Nomis (3:15, 8:53, 9:5)
