+++ NAMESPACE
@@ -12,3 +12,4 @@ importFrom(jsonlite,fromJSON)
 importFrom(readr,read_csv)
 importFrom(tibble,as_tibble)
 importFrom(tibble,enframe)
+importFrom(utils,menu)
