+++ DESCRIPTION
@@ -18,7 +18,8 @@ Imports:
   jsonlite,
   tibble,
   readr,
-  dplyr
+  dplyr,
+  utils
 RoxygenNote: 6.0.1
 Suggests:
   knitr,
