### 2018-04-17 bc33402
fixing spelling of 'options'
