### 2018-04-17 f4cb561
updating documentation and fixing more spelling
