### 2018-04-17 cd6be07
fixing documentation so that .onLoad is not aliased to package man
