### 2018-04-17 #190 @stefaniebutland
Hi @evanodell. Great to hear you're interested in contributing a post about `nomisr`. The format is really up to you, but you could take publication dates into account. Right now we have a full schedule of longer narrative posts up to ~June so if you would like it up quickly, a technote is the way to go. We'll tweet from @rOpenSci either way.

You can read example narrative posts about onboarded packages here: https://ropensci.org/tags/onboarding/ and technotes are here: https://ropensci.org/technotes/

Technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post. The only difference in YAML for technote is
categories:
- technotes

I will add the topicid before publishing.

Happy to answer questions here.
