### 2018-04-17 #190 @karthik
@evanodell Final suggestion. If you found feedback from the reviewers valuable, consider adding them as reviewers in the description of your package (this is entirely up to you and not a requirement).  If you decide to, you'll need:

- The ORCID for both reviewers (if the reviewers don't have it, they should sign up for one here: https://orcid.org/)
- Follow this format: https://github.com/ropensci/rdflib/blob/master/DESCRIPTION#L8-L10
- More context for why such attribution matters: https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/
