### 2018-04-17 #190 @evanodell
@stefaniebutland A technote seems good to me. I can send it to you by the end of the week if that works?
