+++ tests/testthat/test-load.R
@@ -1,0 +1,7 @@
+context("test-load.R")
+
+test_that("package loading works", {
+ 
+  expect_silent(library(nomisr))  
+  
+})
