+++ R/overview.R
@@ -56,7 +56,7 @@
 #' @examples \donttest{
 #'
 #' library(dplyr)
-#' 
+#'
 #' q <- nomis_overview("NM_1650_1")
 #'
 #' q %>% tidyr::unnest(name) %>% glimpse()
