### 2018-04-18 #190 @karthik
@evanodell Thanks again for the submission. I'll close the issue since the review is complete but please feel free to continue the conversation with Stefanie re the blog post.
